% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicwas.R
\name{ctRUV}
\alias{ctRUV}
\title{Remove Unwanted Variations prior to applying ctassoc}
\usage{
ctRUV(X, W, Y)
}
\arguments{
\item{X}{Matrix (or vector) of phenotypes (and covariates); samples x phenotype(s).}

\item{W}{Matrix of proportion of cell types; samples x cell types.}

\item{Y}{Matrix of bulk omics measurements; probes x samples.
X, W, Y should be numeric.}
}
\value{
Y adjusted for the unwanted variations.
First, linear regression of Y for the full model is computed.
The first ten principal components of the residual is defined as
 the unwanted variations, and subtracted from Y.
}
\description{
Remove Unwanted Variations prior to applying ctassoc
}
\seealso{
ctassoc
}
