% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExsmall.R
\docType{data}
\name{GTExsmall}
\alias{GTExsmall}
\title{Small Subset of GTEx Dataset}
\format{
An object of class \code{list} of length 4.
}
\source{
\href{https://gtexportal.org}{GTEx}
}
\usage{
data(GTExsmall)
}
\description{
The dataset includes gene expression measured in whole blood for 389 samples.
A subset of 500 genes were randomly selected from the original dataset.
}
\examples{
data(GTExsmall)
X = GTExsmall$X
Y = GTExsmall$Y + 1
Y = Y[seq(1, 20), ] # for brevity
W = GTExsmall$W
C = GTExsmall$C
result = ctassoc(X, W, Y, C = C)
result$coefficients
}
\seealso{
ctassoc
}
\keyword{datasets}
