% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicwas.R
\name{ctassoc}
\alias{ctassoc}
\title{Cell-Type-Specific Association Testing}
\usage{
ctassoc(X, W, Y, test = "ridge", num.cores = 1, chunk.size = 100)
}
\arguments{
\item{X}{Matrix (or vector) of phenotypes (and covariates); samples x phenotype(s).}

\item{W}{Matrix of proportion of cell types; samples x cell types.}

\item{Y}{Matrix of bulk omics measurements; probes x samples.
X, W, Y should be numeric.}

\item{test}{Statistical test to apply; either ridge, full or marginal.}

\item{num.cores}{Number of CPU cores to use.
Full and marginal tests are run in serial, thus num.cores is ignored.}

\item{chunk.size}{The size of job for a CPU core in one batch.
If you have many cores but limited memory, and there is a memory failure,
decrease this parameter.}
}
\value{
A list with one element, which is named "coefficients".
The element gives the estimate, statistic, p.value in tibble format.
}
\description{
Cell-Type-Specific Association Testing
}
\examples{
data(GSE42861small)
X = GSE42861small$X
Y = GSE42861small$Y
W = GSE42861small$W
Y = ctRUV(X, W, Y)
result = ctassoc(X, W, Y)
result$coefficients

}
\seealso{
ctRUV
}
