% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicwas.R
\name{ctassoc}
\alias{ctassoc}
\title{Cell-Type-Specific Association Testing}
\usage{
ctassoc(X, W, Y, test = "ridge", num.cores = 1, chunk.size = 1000)
}
\arguments{
\item{X}{Matrix (or vector) of phenotypes (and covariates); samples x phenotype(s).}

\item{W}{Matrix of proportion of cell types; samples x cell types.}

\item{Y}{Matrix of bulk omics measurements; probes x samples.
X, W, Y should be numeric.}

\item{test}{Statistical test to apply; either \code{reducedrankridge},
\code{ridge}, \code{full} or \code{marginal}.}

\item{num.cores}{Number of CPU cores to use.
Full and marginal tests are run in serial, thus num.cores is ignored.}

\item{chunk.size}{The size of job for a CPU core in one batch.
If you have many cores but limited memory, and there is a memory failure,
decrease num.cores and/or chunk.size.}
}
\value{
A list with one element, which is named "coefficients".
The element gives the estimate, statistic, p.value in tibble format.
}
\description{
Cell-Type-Specific Association Testing
}
\details{
Let the indexes be
cell type \eqn{h}, sample \eqn{i}, CpG site \eqn{j},
and phenotype (disease state and covariates, such as age, sex) \eqn{k}.
There are three input data, phenotype \eqn{X_i_k},
cell type proportion \eqn{W_i_h} and
methylation level \eqn{Y_j_i}.
For each tissue sample, the cell type proportion \eqn{W_i_h}
is the proportion of each cell type in the bulk tissue,
which is measured or imputed beforehand.
The tissue-level methylation level \eqn{Y_j_i} is measured and provided as input.

The cell-type-specific methylation level \eqn{Z_h_j_i} is not observed
and is treated as a hidden variable.
The parameter we estimate is
the cell-type-specific effect of phenotype \eqn{\beta_h_j_k},
and we also fit the cell-type-specific basal methylation level \eqn{\mu_h_j}.

We assume normal distribution for the cell-type-specific methylation level,
\deqn{Z_h_j_i ~ N(\mu_h_j + \sum_k \beta_h_j_k * X_i_k, \sigma^2_h_j).}
Since the bulk tissue methylation level is the sum weighted by \eqn{W_i_h},
\deqn{Y_j_i ~ N(\sum_h W_i_h {\mu_h_j + \sum_k \beta_h_j_k * X_i_k},
      \tau^2_j + \sum_h W_i_h * \sigma^2_h_j).}
In practice, we replace the error term involving \eqn{\sigma^2_h_j} and
\eqn{\tau^2_j}, simply by \eqn{\tau^2_j}.

The \code{full} model is the linear regression
\deqn{Y_j_i ~ (\sum_h \mu_h_j * W_i_h) + (\sum_h_k \beta_h_j_k * W_i_h * X_i_k) +
 error.}
The \code{ridge} model is the ridge regression for the same equation,
aiming to cope with multicollinearity of the independent variables.
The \code{marginal} model tests the phenotype association only in one
cell type \eqn{h}, under the linear regression,
\deqn{Y_j_i ~ (\sum_h' \mu_h'_j * W_i_h') + (\sum_k \beta_h_j_k * W_i_h * X_i_k) +
 error.}
}
\examples{
data(GSE42861small)
X = GSE42861small$X
Y = GSE42861small$Y
W = GSE42861small$W
Y = ctRUV(X, W, Y)
result = ctassoc(X, W, Y)
result$coefficients

}
\seealso{
ctRUV
}
