% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicwas.R
\name{ctRUV}
\alias{ctRUV}
\title{Remove Unwanted Variations prior to applying ctassoc}
\usage{
ctRUV(X, W, Y, C = NULL, method = "PCA")
}
\arguments{
\item{X}{Matrix (or vector) of phenotypes; samples x phenotype(s).}

\item{W}{Matrix of proportion of cell types; samples x cell types.}

\item{Y}{Matrix of bulk omics measurements; probes x samples.}

\item{C}{Matrix (or vector) of covariates; samples x covariate(s).
X, W, Y, C should be numeric.}

\item{method}{"PCA" or "SVA"}
}
\value{
Y adjusted for the unwanted variations.
}
\description{
Remove Unwanted Variations prior to applying ctassoc
}
\details{
First, for each CpG site, the full linear model of the \code{ctassoc}
function is fitted, and the residual is computed.
For the residuals over all CpG sites, the principal components (PCs)
are computed.
The top ten PCs are regarded as the unwanted variations,
and subtracted from \code{Y}.
}
\seealso{
ctassoc
}
